chore: separate strel functions to submodule
A single structuring_element.jl file becomes inconvinient to
navigate through. Thus I believe it's a good idea to separate it
into a submodule before adding new features.

In the future we might make this a standalone package.
