+++ Project.toml
@@ -5,12 +5,14 @@ version = "0.3.2"
 [deps]
 ImageCore = "a09fc81d-aa75-5fe9-8630-4744c3626534"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 TiledIteration = "06e1c1a7-607b-532d-9fad-de7d9aa2abac"
 
 [compat]
 Documenter = "0.24, 0.25"
 ImageCore = "0.9"
+OffsetArrays = "1.9"
 Requires = "1"
 TiledIteration = "0.3.1"
 julia = "1.6"
